move autocomplete execution to corresponding command file
