+++ commands/doc.js
@@ -1,5 +1,6 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const {renderInter: render} = require("../src/render");
+const {docAutocomplete} = require("../src/mathworks-docs");
 
 module.exports = {
     data: new SlashCommandBuilder()
@@ -15,4 +16,22 @@ module.exports = {
         let docURL = 'https://mathworks.com/help/' + interaction.options.getString('query');
         await render(interaction, 'doc.md', {url: docURL})
     },
+    async autocompleteExecute(client, interaction) {
+        const defaultChoices = [{
+            name:'Getting Started',
+            value:'matlab/getting-started-with-matlab.html'
+        }];
+
+        const focusedValue = interaction.options.getFocused();
+        if (!focusedValue) {
+            await interaction.respond(defaultChoices).catch(console.log)
+            return
+        }
+        const searchResults = await docAutocomplete(focusedValue);
+        if (!searchResults.length) {
+            await interaction.respond(defaultChoices).catch(console.log)
+            return
+        }
+        await interaction.respond(searchResults).catch(console.log)
+    }
 };
