+++ events/autocompleteInteraction.js
@@ -1,23 +1,14 @@
 const {docAutocomplete} = require("../src/mathworks-docs");
 module.exports = {
     async autocompleteExecute(client, interaction) {
-        if (interaction.commandName === "doc") {
-            const defaultChoices = [{
-                name:'Getting Started',
-                value:'matlab/getting-started-with-matlab.html'
-            }];
+        const command = client.commands.get(interaction.commandName);
 
-            const focusedValue = interaction.options.getFocused();
-            if (!focusedValue) {
-                await interaction.respond(defaultChoices).catch(console.log)
-                return
-            }
-            const searchResults = await docAutocomplete(focusedValue);
-            if (!searchResults.length) {
-                await interaction.respond(defaultChoices).catch(console.log)
-                return
-            }
-            await interaction.respond(searchResults).catch(console.log)
+        if (!command) return;
+
+        try {
+            await command.autocompleteExecute(client, interaction);
+        } catch (error) {
+            console.error(error);
         }
     }
 }
